Changed some Matrix3 functions so they use doubles instead of ints. Made the code delete the cube then reapply the points, so changes can be made to the cube via rotations and translations. Added face culling. Winding has to be fixed.

+++ SFMLOpenGL/Matrix3.cpp
@@ -172,7 +172,7 @@ Matrix3 Matrix3::Inverse(Matrix3 M1)
 
 }
 
-Matrix3 Matrix3::Rotation(int _angle)
+Matrix3 Matrix3::Rotation(double _angle)
 {
 	double radians = PI / 180 * _angle;
 	Matrix3 answer;
@@ -189,7 +189,7 @@ Matrix3 Matrix3::Rotation(int _angle)
 	return answer;
 }
 
-Matrix3 Matrix3::Translate(int dx, int dy)
+Matrix3 Matrix3::Translate(double dx, double dy)
 {
 	Matrix3 answer;
 	answer.A11 = 1;
@@ -228,7 +228,7 @@ Matrix3 Matrix3::operator-()
 			A31 * -1, A32 * -1, A33 * -1 };
 }
 
-Matrix3 Matrix3::RotationX(int _angle)
+Matrix3 Matrix3::RotationX(double _angle)
 {
 	double radians = 3.14159 / 180 * _angle;
 	Matrix3 answer;
@@ -245,7 +245,7 @@ Matrix3 Matrix3::RotationX(int _angle)
 	return answer;
 }
 
-Matrix3 Matrix3::RotationY(int _angle)
+Matrix3 Matrix3::RotationY(double _angle)
 {
 	double radians = 3.14159 / 180 * _angle;
 	Matrix3 answer;
@@ -262,7 +262,7 @@ Matrix3 Matrix3::RotationY(int _angle)
 	return answer;
 }
 
-Matrix3 Matrix3::RotationZ(int _angle)
+Matrix3 Matrix3::RotationZ(double _angle)
 {
 	double radians = 3.14159 / 180 * _angle;
 	Matrix3 answer;
